requirements.txt
pymssql
woocommerce
python-dotenv